+++ services/opentsdb/service.go
@@ -83,7 +83,6 @@ func NewService(c Config) (*Service, error) {
 	d := c.WithDefaults()
 
 	s := &Service{
-		done:            make(chan struct{}),
 		tls:             d.TLSEnabled,
 		cert:            d.Certificate,
 		err:             make(chan error),
@@ -106,6 +105,11 @@ func (s *Service) Open() error {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
+	if !s.closed() {
+		return nil // Already open.
+	}
+	s.done = make(chan struct{})
+
 	s.Logger.Println("Starting OpenTSDB service")
 
 	if _, err := s.MetaClient.CreateDatabase(s.Database); err != nil {
@@ -118,7 +122,7 @@ func (s *Service) Open() error {
 
 	// Start processing batches.
 	s.wg.Add(1)
-	go s.processBatches(s.batcher)
+	go func() { defer s.wg.Done(); s.processBatches(s.batcher) }()
 
 	// Open listener.
 	if s.tls {
@@ -149,8 +153,8 @@ func (s *Service) Open() error {
 
 	// Begin listening for connections.
 	s.wg.Add(2)
-	go s.serveHTTP()
-	go s.serve()
+	go func() { defer s.wg.Done(); s.serve() }()
+	go func() { defer s.wg.Done(); s.serveHTTP() }()
 
 	return nil
 }
@@ -160,16 +164,44 @@ func (s *Service) Close() error {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
-	if s.ln != nil {
-		return s.ln.Close()
+	if s.closed() {
+		return nil // Already closed.
 	}
+	close(s.done)
+
+	// Close the listeners.
+	if err := s.ln.Close(); err != nil {
+		return err
+	}
+	if err := s.httpln.Close(); err != nil {
+		return err
+	}
+
+	s.wg.Wait()
+	s.done = nil
 
 	if s.batcher != nil {
 		s.batcher.Stop()
 	}
-	close(s.done)
-	s.wg.Wait()
+
 	return nil
+}
+
+// Closed returns true if the service is currently closed.
+func (s *Service) Closed() bool {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	return s.closed()
+}
+
+func (s *Service) closed() bool {
+	select {
+	case <-s.done:
+		// Service is closing.
+		return true
+	default:
+		return s.done == nil
+	}
 }
 
 // SetLogOutput sets the writer to which all logs are written. It must not be
@@ -225,7 +257,7 @@ func (s *Service) Statistics(tags map[string]string) []models.Statistic {
 }
 
 // Err returns a channel for fatal errors that occur on the listener.
-func (s *Service) Err() <-chan error { return s.err }
+// func (s *Service) Err() <-chan error { return s.err }
 
 // Addr returns the listener's address. Returns nil if listener is closed.
 func (s *Service) Addr() net.Addr {
@@ -237,8 +269,6 @@ func (s *Service) Addr() net.Addr {
 
 // serve serves the handler from the listener.
 func (s *Service) serve() {
-	defer s.wg.Done()
-
 	for {
 		// Wait for next connection.
 		conn, err := s.ln.Accept()
@@ -282,6 +312,7 @@ func (s *Service) handleConn(conn net.Conn) {
 	// Otherwise handle in telnet format.
 	s.wg.Add(1)
 	s.handleTelnetConn(conn)
+	s.wg.Done()
 }
 
 // handleTelnetConn accepts OpenTSDB's telnet protocol.
@@ -289,7 +320,6 @@ func (s *Service) handleConn(conn net.Conn) {
 //   put sys.cpu.user 1356998400 42.5 host=webserver01 cpu=0
 func (s *Service) handleTelnetConn(conn net.Conn) {
 	defer conn.Close()
-	defer s.wg.Done()
 	defer atomic.AddInt64(&s.stats.ActiveTelnetConnections, -1)
 	atomic.AddInt64(&s.stats.ActiveTelnetConnections, 1)
 	atomic.AddInt64(&s.stats.HandledTelnetConnections, 1)
@@ -408,7 +438,6 @@ func (s *Service) serveHTTP() {
 
 // processBatches continually drains the given batcher and writes the batches to the database.
 func (s *Service) processBatches(batcher *tsdb.PointBatcher) {
-	defer s.wg.Done()
 	for {
 		select {
 		case batch := <-batcher.Out():