+++ services/opentsdb/service_test.go
@@ -1,6 +1,7 @@
 package opentsdb_test
 
 import (
+	"fmt"
 	"io/ioutil"
 	"log"
 	"net"
@@ -12,24 +13,62 @@ import (
 	"time"
 
 	"github.com/davecgh/go-spew/spew"
+	"github.com/influxdata/influxdb/internal"
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/services/meta"
 	"github.com/influxdata/influxdb/services/opentsdb"
 )
 
+func Test_Service_OpenClose(t *testing.T) {
+	service := NewService("db0", "127.0.0.1:45362")
+
+	// Closing a closed service is fine.
+	if err := service.Service.Close(); err != nil {
+		t.Fatal(err)
+	}
+
+	// Closing a closed service again is fine.
+	if err := service.Service.Close(); err != nil {
+		t.Fatal(err)
+	}
+
+	if err := service.Service.Open(); err != nil {
+		t.Fatal(err)
+	}
+
+	// Opening an already open service is fine.
+	if err := service.Service.Open(); err != nil {
+		t.Fatal(err)
+	}
+
+	// Reopening a previously opened service is fine.
+	if err := service.Service.Close(); err != nil {
+		t.Fatal(err)
+	}
+
+	if err := service.Service.Open(); err != nil {
+		t.Fatal(err)
+	}
+
+	// Tidy up.
+	if err := service.Service.Close(); err != nil {
+		t.Fatal(err)
+	}
+}
+
 // Ensure a point can be written via the telnet protocol.
 func TestService_Telnet(t *testing.T) {
 	t.Parallel()
 
-	s := NewService("db0")
-	if err := s.Open(); err != nil {
+	s := NewService("db0", "127.0.0.1:0")
+	if err := s.Service.Open(); err != nil {
 		t.Fatal(err)
 	}
-	defer s.Close()
+	defer s.Service.Close()
 
 	// Mock points writer.
 	var called int32
-	s.PointsWriter.WritePointsFn = func(database, retentionPolicy string, consistencyLevel models.ConsistencyLevel, points []models.Point) error {
+	s.WritePointsFn = func(database, retentionPolicy string, consistencyLevel models.ConsistencyLevel, points []models.Point) error {
 		atomic.StoreInt32(&called, 1)
 
 		if database != "db0" {
@@ -50,7 +89,7 @@ func TestService_Telnet(t *testing.T) {
 	}
 
 	// Open connection to the service.
-	conn, err := net.Dial("tcp", s.Addr().String())
+	conn, err := net.Dial("tcp", s.Service.Addr().String())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -84,15 +123,15 @@ func TestService_Telnet(t *testing.T) {
 func TestService_HTTP(t *testing.T) {
 	t.Parallel()
 
-	s := NewService("db0")
-	if err := s.Open(); err != nil {
+	s := NewService("db0", "127.0.0.1:0")
+	if err := s.Service.Open(); err != nil {
 		t.Fatal(err)
 	}
-	defer s.Close()
+	defer s.Service.Close()
 
 	// Mock points writer.
 	var called bool
-	s.PointsWriter.WritePointsFn = func(database, retentionPolicy string, consistencyLevel models.ConsistencyLevel, points []models.Point) error {
+	s.WritePointsFn = func(database, retentionPolicy string, consistencyLevel models.ConsistencyLevel, points []models.Point) error {
 		called = true
 		if database != "db0" {
 			t.Fatalf("unexpected database: %s", database)
@@ -113,7 +152,7 @@ func TestService_HTTP(t *testing.T) {
 	}
 
 	// Write HTTP request to server.
-	resp, err := http.Post("http://"+s.Addr().String()+"/api/put", "application/json", strings.NewReader(`{"metric":"sys.cpu.nice", "timestamp":1346846400, "value":18, "tags":{"host":"web01", "dc":"lga"}}`))
+	resp, err := http.Post("http://"+s.Service.Addr().String()+"/api/put", "application/json", strings.NewReader(`{"metric":"sys.cpu.nice", "timestamp":1346846400, "value":18, "tags":{"host":"web01", "dc":"lga"}}`))
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -131,40 +170,44 @@ func TestService_HTTP(t *testing.T) {
 }
 
 type Service struct {
-	*opentsdb.Service
-	PointsWriter PointsWriter
-}
-
-// NewService returns a new instance of Service.
-func NewService(database string) *Service {
-	srv, _ := opentsdb.NewService(opentsdb.Config{
-		BindAddress:      "127.0.0.1:0",
-		Database:         database,
-		ConsistencyLevel: "one",
-	})
-	s := &Service{Service: srv}
-	s.Service.PointsWriter = &s.PointsWriter
-	s.Service.MetaClient = &DatabaseCreator{}
-
-	if !testing.Verbose() {
-		s.Logger = log.New(ioutil.Discard, "", log.LstdFlags)
-	}
-
-	return s
-}
-
-// PointsWriter represents a mock impl of PointsWriter.
-type PointsWriter struct {
+	Service       *opentsdb.Service
+	MetaClient    *internal.MetaClientMock
 	WritePointsFn func(database, retentionPolicy string, consistencyLevel models.ConsistencyLevel, points []models.Point) error
 }
 
-func (w *PointsWriter) WritePoints(database, retentionPolicy string, consistencyLevel models.ConsistencyLevel, points []models.Point) error {
-	return w.WritePointsFn(database, retentionPolicy, consistencyLevel, points)
+// NewService returns a new instance of Service.
+func NewService(database string, bind string) *Service {
+	s, err := opentsdb.NewService(opentsdb.Config{
+		BindAddress:      bind,
+		Database:         database,
+		ConsistencyLevel: "one",
+	})
+
+	if err != nil {
+		panic(err)
+	}
+
+	service := &Service{
+		Service:    s,
+		MetaClient: &internal.MetaClientMock{},
+	}
+
+	service.MetaClient.CreateDatabaseFn = func(db string) (*meta.DatabaseInfo, error) {
+		if got, exp := db, database; got != exp {
+			return nil, fmt.Errorf("got %v, expected %v", got, exp)
+		}
+		return nil, nil
+	}
+
+	if !testing.Verbose() {
+		service.Service.Logger = log.New(ioutil.Discard, "", log.LstdFlags)
+	}
+
+	service.Service.MetaClient = service.MetaClient
+	service.Service.PointsWriter = service
+	return service
 }
 
-type DatabaseCreator struct {
-}
-
-func (d *DatabaseCreator) CreateDatabase(name string) (*meta.DatabaseInfo, error) {
-	return nil, nil
+func (s *Service) WritePoints(database, retentionPolicy string, consistencyLevel models.ConsistencyLevel, points []models.Point) error {
+	return s.WritePointsFn(database, retentionPolicy, consistencyLevel, points)
 }